# RealTimeChatApp-FE/src/app/service/models/chat-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ChatResponse {
  chatName?: string;
  id?: string;
  lastMessage?: string;
  lastMessageTime?: string;
  recipientId?: string;
  recipientOnline?: boolean;
  senderId?: string;
  unreadCount?: number;
}